requirements.txt
Django==1.11.3
django-appconf==1.0.2
django-avatar==4.0.0
django-ckeditor==5.3.0
django-constance==2.0.0
django-countries==4.6.1
django-js-asset==0.1.1
django-picklefield==1.0.0
olefile==0.44
Pillow==4.2.1
pytz==2017.2
